desktop/requirements.txt
customtkinter>=5.2.0
pillow>=10.0.0
requests>=2.31.0
opencv-python>=4.8.0
numpy>=1.24.0
